fix: мелкие фиксы

+++ src/main/java/ru/yandex/practicum/filmorate/storage/feed/FeedStorage.java
@@ -5,5 +5,5 @@ import ru.yandex.practicum.filmorate.model.FeedEvent;
 import java.util.List;
 
 public interface FeedStorage {
-    List<FeedEvent> getUserFeed(int userId);
+    List<FeedEvent> getUserFeed(long userId);
 }
